+++ order-service/src/main/java/ru/skillbox/orderservice/domain/OrderDto.java
@@ -13,5 +13,5 @@ public class OrderDto {
 
     private String destinationAddress;
 
-    private Long cost;
+    private Double cost;
 }
